+++ zetaclient/config/types.go
@@ -4,13 +4,16 @@ import (
 	"encoding/json"
 	"fmt"
 	"maps"
+	"net"
 	"os"
+	"sort"
 	"strings"
 	"sync"
 
 	"github.com/asaskevich/govalidator"
 	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/pkg/errors"
+	"github.com/rs/zerolog"
 	"github.com/showa-93/go-mask"
 
 	"github.com/zeta-chain/node/pkg/chains"
@@ -159,10 +162,8 @@ func (c Config) Validate() error {
 		return errors.Errorf("reason: invalid public IP, got: %s", c.PublicIP)
 	}
 
-	// TODO: add back public DNS check once libp2p-go can handle public DNS
-	// https://github.com/zeta-chain/node/issues/4433
-	if c.PublicDNS != "" {
-		return errors.Errorf("reason: public DNS is not supported, got: %s", c.PublicDNS)
+	if c.PublicDNS != "" && !govalidator.IsDNSName(c.PublicDNS) {
+		return errors.Errorf("reason: invalid public DNS, got: %s", c.PublicDNS)
 	}
 
 	if _, err := chains.ZetaChainFromCosmosChainID(c.ChainID); err != nil {
@@ -218,6 +219,45 @@ func (c Config) Validate() error {
 	}
 
 	return nil
+}
+
+// ResolvePublicDNS4 resolves the public DNS to an IPv4 address when public IP isn't set.
+// For simplicity, the 1st resolved IPv4 address is used if multiple IP addresses are found.
+func (c Config) ResolvePublicIP(logger zerolog.Logger) (string, error) {
+	// return public IP if already set
+	if c.PublicIP != "" {
+		return c.PublicIP, nil
+	}
+
+	// return error if public DNS isn't set
+	if c.PublicDNS == "" {
+		return "", errors.New("no public IP or DNS is provided")
+	}
+
+	// lookup IP addresses
+	ips, err := net.LookupIP(c.PublicDNS)
+	if err != nil {
+		return "", errors.Wrapf(err, "unable to resolve IP addresses for public DNS %q", c.PublicDNS)
+	} else if len(ips) == 0 {
+		return "", fmt.Errorf("no IP address resolved for public DNS %q", c.PublicDNS)
+	}
+
+	// sort IP addresses to be deterministic
+	strIPs := make([]string, len(ips))
+	for i, ip := range ips {
+		strIPs[i] = ip.String()
+	}
+	sort.Strings(strIPs)
+
+	// go-tss requires a valid IPv4 address
+	for _, ip := range strIPs {
+		if govalidator.IsIPv4(ip) {
+			logger.Info().Str("ip", ip).Msg("resolved public IP")
+			return ip, nil
+		}
+	}
+
+	return "", fmt.Errorf("no IPv4 address resolved for public DNS %q", c.PublicDNS)
 }
 
 // GetEVMConfig returns the EVM config for the given chain ID